:sparkles: Add InputContainer
