+++ src/App.jsx
@@ -1,36 +1,12 @@
-import { useDispatch, useSelector } from 'react-redux';
-
-import Page from './Page';
-
-import {
-  updateTaskTitle,
-  addTask,
-} from './actions';
-
-function selector(state) {
-  return {
-    taskTitle: state.taskTitle,
-  };
-}
+import InputContainer from './InputContainer';
+import ListContainer from './ListContainer';
 
 export default function App() {
-  const dispatch = useDispatch();
-
-  function handleChangeTitle(event) {
-    dispatch(updateTaskTitle(event.target.value));
-  }
-
-  function handleClickAddTask() {
-    dispatch(addTask());
-  }
-
-  const { taskTitle } = useSelector(selector);
-
   return (
-    <Page
-      taskTitle={taskTitle}
-      onChangeTitle={handleChangeTitle}
-      onClickAddTask={handleClickAddTask}
-    />
+    <div>
+      <h1>To-do</h1>
+      <InputContainer />
+      <ListContainer />
+    </div>
   );
 }
