+++ src/InputContainer.test.jsx
@@ -1,0 +1,31 @@
+import { render, fireEvent } from '@testing-library/react';
+
+import { useDispatch, useSelector } from 'react-redux';
+
+import InputContainer from './InputContainer';
+
+jest.mock('react-redux');
+test('InputContainer', () => {
+  const dispatch = jest.fn();
+
+  useDispatch.mockImplementation(() => dispatch);
+
+  useSelector.mockImplementation((selector) => selector({
+    taskTitle: 'New Title',
+  }));
+
+  const { getByText, getByDisplayValue } = render((
+    <InputContainer />
+  ));
+
+  expect(getByText(/추가/)).not.toBeNull();
+  expect(getByDisplayValue(/New Title/)).not.toBeNull();
+
+  fireEvent.click(getByText(/추가/));
+
+  expect(dispatch).toBeCalledWith({ type: 'addTask' });
+  // expect(getByText(/New Title/)).not.toBeNull();
+
+  // TODO: 통합 테스트 코드 작성
+  // CodeceptJS => 실제 브라우저에서 사용자 테스트 실행 가능.
+});
